+++ rocAL/include/augmentations/augmentations_nodes.h
@@ -55,3 +55,4 @@ THE SOFTWARE.
 #include "augmentations/node_copy.h"
 #include "augmentations/node_nop.h"
 #include "augmentations/node_sequence_rearrange.h"
+#include "augmentations/audio_augmentations/node_preemphasis_filter.h"